+++ DUBUFrameWork/InternalClient.cpp
@@ -269,7 +269,7 @@ bool DUBU::InternalClient::RecvDispatch(Uint8* buffer, Uint16 size)
             if (header->sequenceNo_ != rps.recvRepeatSeq_ + 1)
             {
                 // 캐싱
-                Uint64 del = header->sequenceNo_ - rps.recvRepeatSeq_;
+                Uint32 del = header->sequenceNo_ - rps.recvRepeatSeq_;
                 if (del >= DEFAULT_WINDOW_COUNT)
                 {
                     // 최대 캐싱 가능크기는 64넘기면 캐싱안하고 넘어간다.
@@ -343,7 +343,7 @@ bool DUBU::InternalClient::RecvDispatch(Uint8* buffer, Uint16 size)
             if (header->sequenceNo_ != rpsNo_.recvRepeatSeq_ + 1)
             {
                 // 현재 순서가 아닌 패킷인 경우
-                Uint64 del = header->sequenceNo_ - rpsNo_.recvRepeatSeq_;
+                Uint32 del = header->sequenceNo_ - rpsNo_.recvRepeatSeq_;
 
                 if (del >= DEFAULT_WINDOW_COUNT)
                 {
@@ -412,7 +412,7 @@ bool DUBU::InternalClient::RecvDispatch(Uint8* buffer, Uint16 size)
     }
     else
     {
-        Uint64 del = header->sequenceNo_ - rNopsNo_.recvRepeatSeq_;
+        Uint32 del = header->sequenceNo_ - rNopsNo_.recvRepeatSeq_;
         if (del == 0 || del >= NO_REPEAT_ACCEPT_RANGE)
         {
             // TODO : NO_REPEAT_ACCEPT_RANGE <- 이거 넘어가면 문제있는거다.
@@ -703,7 +703,7 @@ void DUBU::InternalClient::RecvChunckPacket(Uint8* buffer, Uint16 size)
     Uint8 chunck = header->flags_ & Packet::PacketHeaderFlag::CHUNK;
     if (chunck == Packet::PacketHeaderFlag::CHUNK)
     {
-        Uint64 seqNo = header->sequenceNo_;
+        Uint32 seqNo = header->sequenceNo_;
         // 일단 chunk는 채널 사용x : 현재 시스템상 채널은 순서대로 무조건 받는다. 그럼으로 제외
         Packet::ChunkInfo chunckInfo = header->chunkInfo_;
         ChunkPacket& chunckPacket = chunckPakcetInj_.Update(seqNo, chunckInfo.flag_);
